refactor : 운동기록 변경, 수행기록 세트 변경

+++ src/main/java/com/pumping/domain/routineexercise/dto/ExerciseSetResponse.java
@@ -1,4 +1,4 @@
 package com.pumping.domain.routineexercise.dto;
 
-public record ExerciseSetResponse(Integer setCount, Integer weight, Integer repetition) {
+public record ExerciseSetResponse(Integer setCount, Float weight, Integer repetition) {
 }
